+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- AppImage is able to create virtual python environments to be used to develop plugins
+
+
 ## [4.1.1] - 2023-11-01
 
 ### Fixed